# demo-api/openapi.yaml
openapi: 3.0.0
info:
  title: RICO Security Testing Playground
  description: |
    Intentionally vulnerable API for testing RICO's security scanning capabilities.
    
    ⚠️ WARNING: This API contains intentional security vulnerabilities.
    DO NOT deploy this in production environments.
    
    ## Vulnerability Categories
    - SQL Injection (error-based, boolean-blind, time-based)
    - IDOR (Insecure Direct Object Reference)
    - Missing Authentication
    - SSRF (Server-Side Request Forgery)
    - Path Traversal
    - Command Injection
    - JWT Tampering (signature bypass, alg=none)
    - Business Logic Flaws (negative transfers, role injection)
    
    ## Testing Credentials
    - Regular User: alice / password123
    - Regular User: bob / password456
    - Admin User: admin / admin123
  version: 1.0.0
  contact:
    name: RICO Security
    url: https://github.com/yourusername/rico
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /users/{user_id}:
    get:
      tags: [Users]
      summary: Get user by ID (VULNERABLE - IDOR)
      description: |
        🔴 VULNERABLE: No authorization check - any user can access any user's data.
        Test IDOR by accessing other users' IDs.
        
        Examples:
        - GET /users/1 → alice's data
        - GET /users/2 → bob's data
        - GET /users/3 → admin's data
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: User data
        '404':
          description: User not found
...
  /users/{user_id}/orders:
    get:
      tags: [Users]
      summary: Get user orders (VULNERABLE - IDOR)
      description: |
        🔴 VULNERABLE: No authorization - access any user's orders.
        
        Examples:
        - GET /users/1/orders → alice's orders
        - GET /users/2/orders → bob's orders
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User orders
  /users/{user_id}/orders/secure:
    get:
      tags: [Users]
      summary: Get user orders (SECURE)
      description: |
        🟢 SECURE: Proper authorization - users can only access their own orders.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User orders
        '403':
          description: Forbidden
  /sqli/search:
    get:
      tags: [SQL Injection Tests]
      summary: Search users (VULNERABLE - Error-based & Boolean-blind SQLi)
      description: |
        🔴 VULNERABLE: Direct SQL injection in search parameter.
        
        Supports multiple SQLi detection methods:
        - Error-based: Returns SQL errors
        - Boolean-blind: Different responses for TRUE/FALSE
        - Response inference: Content changes based on query
        
        Test payloads:
        - `name=alice` (normal search)
        - `name=alice' OR '1'='1` (TRUE - returns all users)
        - `name=alice' AND '1'='2` (FALSE - returns empty)
        - `name=alice' AND 1=1--` (TRUE condition)
        - `name=alice' AND 1=2--` (FALSE condition)
        - `name=' UNION SELECT * FROM users--` (SQL error)
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          example: alice
      responses:
        '200':
          description: Search results
        '500':
          description: SQL error (injection detected)
  /sqli/search/secure:
    get:
      tags: [SQL Injection Tests]
      summary: Search users (SECURE - Parameterized)
      description: |
        🟢 SECURE: Uses parameterized queries and input validation.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT